python/requirements.txt
dnspython
requests
pyasn
beautifulsoup4
lxml
publicsuffix2
